@@ -5,7 +5,7 @@ Get consecutively unique elements from an array.
 
 @example
 ```
-import uniqueRandomArray = require('unique-random-array');
+import uniqueRandomArray from 'unique-random-array';
 
 const random = uniqueRandomArray([1, 2, 3, 4]);
 
@@ -13,8 +13,4 @@ console.log(random(), random(), random(), random());
 //=> 4 2 1 4
 ```
 */
-declare function uniqueRandomArray<ValueType>(
-	array: readonly ValueType[]
-): () => ValueType;
-
-export = uniqueRandomArray;
+export default function uniqueRandomArray<T>(array: readonly T[]): () => T;
