@@ -1,7 +1,6 @@
-'use strict';
-const uniqueRandom = require('unique-random');
+import uniqueRandom from 'unique-random';
 
-module.exports = array => {
+export default function uniqueRandomArray(array) {
 	const random = uniqueRandom(0, array.length - 1);
 	return () => array[random()];
-};
+}
